requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.23
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
elasticsearch==8.10.1
nltk==3.8.1
python-multipart==0.0.6
typing-extensions==4.8.0
requests==2.31.0
google-generativeai==0.3.1
faker==18.13.0
bson==0.5.10
